+++ ms-azure-api/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to the
 
 ## Unreleased
 
+## 0.3.1.0
+
 MSAzureAPI.StorageServices.FileService : add listDirectoriesAndFilesC (stream all response pages from listDirectoriesAndFiles)
 
 ## 0.3.0.0